Keep admission classification within budget
